breadcrumbs and current page logic with useCurrentPage
