+++ app/routes/_app.tsx
@@ -9,36 +9,47 @@ import {
 	BreadcrumbPage,
 	BreadcrumbSeparator,
 } from "@/components/ui/breadcrumb";
+import { useCurrentPage } from "@/hooks/use-current-page";
 import type { SupabaseOutletContext } from "@/lib/supabase/supabase";
-import { Link, Outlet, useOutletContext } from "@remix-run/react";
+import { Link, Outlet, useNavigate, useOutletContext } from "@remix-run/react";
+import { Fragment } from "react";
 
 export default function AuthLayout() {
 	const { supabase, isLoading, user } = useOutletContext<SupabaseOutletContext>();
-
 	if (isLoading || !user) {
 		return <PageLoading />;
 	}
 
+	const { breadcrumbs, activePage } = useCurrentPage();
+
 	return (
-		<>
-			<AdminPanelLayout>
-				<ContentLayout title="Dashboard">
-					<Breadcrumb>
-						<BreadcrumbList>
-							<BreadcrumbItem>
-								<BreadcrumbLink asChild>
-									<Link to="/">Home</Link>
-								</BreadcrumbLink>
-							</BreadcrumbItem>
-							<BreadcrumbSeparator />
-							<BreadcrumbItem>
-								<BreadcrumbPage>Dashboard</BreadcrumbPage>
-							</BreadcrumbItem>
-						</BreadcrumbList>
-					</Breadcrumb>
-					<Outlet context={{ supabase }} />
-				</ContentLayout>
-			</AdminPanelLayout>
-		</>
+		<AdminPanelLayout>
+			<ContentLayout title="Dashboard">
+				<Breadcrumb>
+					<BreadcrumbList>
+						<BreadcrumbItem>
+							<BreadcrumbLink asChild>
+								<Link to="/">Home</Link>
+							</BreadcrumbLink>
+						</BreadcrumbItem>
+						{breadcrumbs?.map(({ href, label }) => (
+							<Fragment key={label}>
+								<BreadcrumbSeparator />
+								<BreadcrumbItem>
+									{label !== activePage?.label ? (
+										<BreadcrumbLink asChild>
+											<Link to={href}>{label}</Link>
+										</BreadcrumbLink>
+									) : (
+										<BreadcrumbPage>{label}</BreadcrumbPage>
+									)}
+								</BreadcrumbItem>
+							</Fragment>
+						))}
+					</BreadcrumbList>
+				</Breadcrumb>
+				<Outlet context={{ supabase }} />
+			</ContentLayout>
+		</AdminPanelLayout>
 	);
 }
